AG-17107 Move `getDataSelectionState` to ag-charts-enterprise
