+++ packages/ag-charts-community/src/chart/data/dataSelectionService.ts
@@ -1,7 +1,8 @@
-import type { DataSet } from './dataSet';
+import type { ISeries, ISeriesProperties, SelectionState, SeriesNodeDatum } from '../series/seriesTypes';
 
-type SeriesLike = { id: string; data: DataSet<unknown> | undefined };
+type SeriesLike = ISeries<number, SeriesNodeDatum<number>, ISeriesProperties, unknown>;
 
 export interface DataSelectionService {
     getSeriesSelectedCount(series: SeriesLike): number;
+    getDataSelectionState(series: SeriesLike, datumIndex: number | undefined): SelectionState | undefined;
 }
